Use cluster DNS to resolve fake external targets
.github/actions/kind-external-targets/action.sh patches the cluster-wide
DNS to be able to resolve the hostnames of the fake external targets in
the cilium TLD.

Some LVH-based workflows use node-local DNS that doesn't know about the
cilium TLD, therefore, it'll fail to resolve the fake external targets
once they are enabled for LVH-based workflows.

Patch the node-local DNS config to forward requests in the cilium TLD to
the cluster-wide DNS server, which is capable of resolving the hostnames
of the fake external targets.

Signed-off-by: Maxim Mikityanskiy <maxim@isovalent.com>